requirements.txt
Flask==3.0.3
google-genai>=0.2.0
python-dotenv==1.0.1
